Feature/gabhijit izel 149 floating ips revisit

+++ package/cloudshell/cp/openstack/command/operations/hidden_operation.py
@@ -1,6 +1,7 @@
 
 class HiddenOperation(object):
-    def __init__(self, instance_service):
+    def __init__(self, instance_service, network_service):
+        self.network_service = network_service
         self.instance_service = instance_service
 
     def delete_instance(self, openstack_session, deployed_app_resource, floating_ip, logger):
@@ -14,7 +15,25 @@ class HiddenOperation(object):
         :rtype None:
         """
         instance_id = deployed_app_resource.vmdetails.uid
-        self.instance_service.terminate_instance(openstack_session=openstack_session,
+
+        instance = self.instance_service.get_instance_from_instance_id(openstack_session=openstack_session,
+                                                                       instance_id=instance_id,
+                                                                       logger=logger)
+
+        if floating_ip:
+            if instance:
+                self.instance_service.detach_floating_ip(openstack_session=openstack_session,
+                                                         floating_ip=floating_ip,
+                                                         instance=instance,
+                                                         logger=logger)
+
+            self.network_service.delete_floating_ip(openstack_session=openstack_session,
+                                                    floating_ip=floating_ip,
+                                                    logger=logger)
+
+        if instance:
+            self.instance_service.terminate_instance(openstack_session=openstack_session,
                                                  instance_id=instance_id,
-                                                 floating_ip=floating_ip,
                                                  logger=logger)
+        else:
+            logger.info("Instance with Instance ID Not found {}. Not deleting".format(instance_id))
